+++ NEWS.md
@@ -5,6 +5,7 @@
 * Improvements to vignette and online help.
 * Added argument `violin.params` to `DurgaPlot` to provide additional control over violin appearance.
 * Added arguments `ef.size.lty` and `ef.size.lwd` to `DurgaPlot` to provide control over effect size error bar appearance. 
+* Added argument `points.adjust` to `DurgaPlot` to allow control over point layout.
 * Bug fix in `DurgaPlot`: `violin.fill = FALSE` now correctly does not fill violins, rather than white fill.
 * Bug fix in `DurgaBrackets`; drawing many brackets sometimes resulted in overlapping brackets.
 * Extend tick marks on right axis to cover CI when plotting standardised effect size on the right.